Move CreateResume into specific folder
