--- src/CVPZ.Application/Resume/UpdateResume.cs
@@ -1,9 +1,0 @@
-﻿using MediatR;
-
-namespace CVPZ.Application.Resume
-{
-    public class UpdateResume : IRequest<UpdateResumeResponse>
-    {
-
-    }
-}
