--- src/CVPZ.Application/Resume/UpdateResumeResponse.cs
@@ -1,7 +1,0 @@
-﻿namespace CVPZ.Application.Resume
-{
-    public class UpdateResumeResponse
-    {
-
-    }
-}
